Use generated API types directly, drop hand-written duplicates (#55)
## Summary

- Delete the hand-written domain types (`Bug`, `Repo`, `UserInfo`, `BugCloseState`, ID newtypes, etc.) and use the progenitor-generated types directly
- `types.rs` is now just re-exports, type aliases, and trait impls (`ValueEnum`, `Formattable`) on the generated types
- No more JSON round-trip conversion layer — client methods return generated types directly
- User-facing CLI text still says "close" (subcommand name, labels, prompts)

Net -120 lines.

## Test plan

- [x] `cargo build` — no warnings
- [x] `cargo test` — passes
- [x] `cargo clippy -- -D warnings` — clean
- [x] `cargo fmt --check` — clean
- [x] `make check` — help docs unchanged

+++ src/commands/repos.rs
@@ -63,7 +63,13 @@ pub async fn handle(command: &RepoCommands, cli: &crate::Cli) -> Result<()> {
                     term.write_line(&format!("Page: {} of {}", page, total_pages))?;
                     Ok(())
                 }
-                _ => crate::output::output_list(&repos.repos, repos.total, *page, *limit, format),
+                _ => crate::output::output_list(
+                    &repos.repos,
+                    repos.total as usize,
+                    *page,
+                    *limit,
+                    format,
+                ),
             }
         }
     }
